Move value converter spec test to properly class
